Created automation scripts in bash
Created a folder called Packaged-Automation that will be scp'd into the remote machine.
This folder has all the config files and scripts to set up the machines.
The script to prepare the remote machine to be ssh'd into is complete.
The script to prepare the master ansible machine is complete.

Ansible is able to ping to all machines.
SSH is working on all machines.
UFW is working as a firewall on all machines.

To Do:
Laravel configuration file needs to be put into the config files directory.
Testing is required for the script that sets up lamp stack with laravel.

Signed-off-by: H1ddenGames <skarim3000@gmail.com>