Updated the scripts and the names of each script to reflect the final nature of the changes.
Changed ansiblemaster.sh to master-ansible-install.sh
Changed lampremote.sh to remote-lamp-install.sh
Changed prepareremote.sh to remote-prepare.sh

Changed the layout of the scripts to be more human readable.
Updated some tasks in the scripts to be in a more logical area of the script.
Added some tasks that were missing in the scripts.

To do:
More testing is required to check if the changes are ready for production.


Signed-off-by: Shahid Karim <skarim3000@gmail.com>